refactor(character): move Class.toFormattedString to CharacterFormatExt
SpellFormatExt.kt no longer owns Character-related display logic.
